requirements.txt
Django==4.2.6
djangorestframework
django-cors-headers
